service file to get images from dalle and new gradle dependency
this service can get images in base64 or url format and returns the string
sync the gradle dependencies since we use a new dependency for API call
call function getImageFromDALLE("a cat sitting on a bed") without the
prompt as the parameter. will return base64 encoded image (or url string)
Right now we need to provide an API Key to check if it works, but it should
work out of the box.
To render images in React @moritzmohn, @xueswang use the following for either base64 or url: https://stackoverflow.com/a/42399865/3647970