Adding new method to set just a single player score
From the UI, one player has no information about the other players,
so when the votes are sent, the UI will just send a PUT request for the
particular user. Therefore the `setPlayerScores` method is not really usable.
Therefore I added a method where a particular user is filtered and then
the score is updated.